+++ src/hooks/Product/Product.hook.ts
@@ -1,54 +1,69 @@
-import { useState } from "react";
 import { Product } from "../../pages/Product/Product.static";
-import { token } from "../../static/token";
 import { endpoint } from "../../static/endpoints/Endpoint";
+import { useMutation, useQuery } from "react-query";
 
 const useProduct = () => {
-  const [products, setProducts] = useState<Product[]>([]);
+  const {
+    data: products,
+    isLoading,
+    error,
+    refetch,
+  } = useQuery("products", getProduct);
 
-  const getProduct = () => {
-    fetch(endpoint.product, {
+  const deleteProductMutation = useMutation(deleteProduct);
+
+  async function getProduct() {
+    const token = localStorage.getItem("accessToken");
+
+    if (!token) {
+      throw new Error("Access token not found");
+    }
+    const res = await fetch(endpoint.product, {
       method: "GET",
       headers: {
         "Content-type": "application/json",
         Authorization: `Bearer ${token}`,
       },
-    })
-      .then((res) => res.json())
-      .then((data: Product[]) => {
-        setProducts(() =>
-          data.map((product) => ({
-            ...product,
-            createdAt: new Date(product.createdAt).toLocaleString(),
-            updatedAt: new Date(product.updatedAt).toLocaleString(),
-          }))
-        );
-      });
-  };
-  const handleDelete = (id: string) => {
-    fetch(endpoint.product + "/soft-delete/" + id, {
+    });
+    if (!res.ok) {
+      throw new Error(`Failed to fetch data: ${res.statusText}`);
+    }
+
+    const data = await res.json();
+
+    if (!Array.isArray(data)) {
+      throw new Error(`Unexpected data format: ${JSON.stringify(data)}`);
+    }
+
+    return data.map((product: Product) => ({
+      ...product,
+      createdAt: new Date(product.createdAt),
+      updatedAt: new Date(product.updatedAt),
+    }));
+  }
+
+  async function deleteProduct(id: string) {
+    const token = localStorage.getItem("accessToken");
+
+    if (!token) {
+      throw new Error("Access token not found");
+    }
+
+    const res = await fetch(`${endpoint.product}/soft-delete/${id}`, {
       method: "DELETE",
       headers: {
         "Content-type": "application/json",
         Authorization: `Bearer ${token}`,
       },
-    })
-    .then((res) => {
-      if (!res.ok) {
-        throw new Error(`Failed to delete product. Status: ${res.status}`);
-      }
-      return res.json();
-    })
-    .then(() => {
-      setProducts((prevProduct) =>
-        prevProduct.filter((product) => product.id !== id)
-      );
-    })
-    .catch((error) => {
-      console.error("Error deleting product:", error);
     });
-  };
-  return { products, getProduct, handleDelete };
+
+    if (!res.ok) {
+      throw new Error(`Failed to delete product: ${res.statusText}`);
+    }
+
+    refetch();
+  }
+  return { products, isLoading, error, deleteProduct: deleteProductMutation.mutate};
 };
 
 export default useProduct;
